+++ src/cico_one_pass.jl
@@ -10,9 +10,10 @@ function get_right_endpoint(
         ),
     scan_bound::Float64 = 9.0,
     scan_tol::Float64 = 1e-3,
-    loss_tol::Float64 = 1e-3, # i do not know how to use it
-    # good results in :LN_NELDERMEAD, :LN_COBYLA, :LN_PRAXIS,
-    # errors in :LN_BOBYQA, :LN_SBPLX, :LN_NEWUOA
+    loss_tol::Float64 = 1e-3, # we have no ideas how to implement loss_tol
+    # better results for LN_NELDERMEAD, :LD_MMA, :LD_SLSQP, :LD_CCSAQ
+    # worse results for :LN_COBYLA, :LN_PRAXIS
+    # errors for :LN_BOBYQA, :LN_SBPLX, :LN_NEWUOA
     local_alg::Symbol = :LN_NELDERMEAD,
     # options for local fitter :max_iter
     max_iter::Int = 10^5,
@@ -23,10 +24,6 @@ function get_right_endpoint(
     n_theta = length(theta_init)
 
     # checking arguments
-    # methods which are not supported
-    if local_alg in [:LN_BOBYQA, :LN_SBPLX, :LN_NEWUOA]
-        @warn "Using local_alg = :"*String(local_alg)*" may result in wrong output."
-    end
     # when using :LN_NELDERMEAD initial parameters should not be zero
     if local_alg == :LN_NELDERMEAD
         zeroParameter = [ isapprox(theta_init[i], 0., atol=1e-2) for i in 1:n_theta]
@@ -40,8 +37,9 @@ function get_right_endpoint(
     local_opt = Opt(local_alg, n_theta)
     ftol_abs!(local_opt, scan_tol) #ftol_abs
 
-    # Constraints function
+    # flags to analyze fitting stop
     out_of_bound::Bool = false
+    #constraints_func_error = false
 
     function constraints_func(x, g) # testing grad methods
     #function constraints_func(x) # testing grad methods    
@@ -53,13 +51,11 @@ function get_right_endpoint(
             @warn "Error when call loss_func($x)"
             throw(e)
         end
-        #println("constr")
-        #@show (x,g)
-        #@show (loss)
-        #@show scan_bound
+        
         if (loss < 0.) && (scan_func(x) > scan_bound)
             out_of_bound = true
             throw(NLopt.ForcedStop())
+            # throw(OutOfBoundException())
         #elseif isapprox(loss, 0., atol=loss_tol)
             #@warn "loss_tol reached... but..."
             #return loss
@@ -76,8 +72,6 @@ function get_right_endpoint(
     end
 
     function obj_func(x,g)
-        #println("obj")
-        #@show (x,g)
         if length(g) > 0
             try ForwardDiff.gradient!(g, scan_func, x)
             catch e
@@ -97,7 +91,6 @@ function get_right_endpoint(
     max_objective!(
         opt,
         obj_func
-        #NLoptAdapter(scan_func,theta_init)
         )
         
     local_optimizer!(opt, local_opt)
@@ -106,13 +99,16 @@ function get_right_endpoint(
     # inequality constraints
     inequality_constraint!(
         opt,
-        #NLoptAdapter(constraints_func,theta_init),
         constraints_func,
         loss_tol
     )
+
+    # version 1: internal :LN_AUGLAG box constrains
     opt.lower_bounds = [tb[1] for tb in theta_bounds]
     opt.upper_bounds = [tb[2] for tb in theta_bounds]
-    #=
+
+    # version 2: creating lower and upper bounds as general constraints
+    #= 
     [ inequality_constraint!(
         opt,
         (x, g) -> x[i] - theta_bounds[i][2],
@@ -123,17 +119,18 @@ function get_right_endpoint(
         (x, g) -> theta_bounds[i][1] - x[i],
         0.
     ) for i in 1:n_theta ]
-=#
+    =#
+
     # start optimization
     (optf, optx, ret) = optimize(opt, theta_init)
 
-    if (ret == :FORCED_STOP && !out_of_bound)
+    if ret == :FORCED_STOP && !out_of_bound
         pp = ProfilePoint[]
         res = (nothing, pp, :LOSS_ERROR_STOP)
     elseif ret == :MAXEVAL_REACHED
         pp = ProfilePoint[]
         res = (nothing, pp, :MAX_ITER_STOP)
-    elseif (ret == :FORCED_STOP && out_of_bound) # successfull result
+    elseif (ret == :FORCED_STOP || ret == :FAILURE) && out_of_bound # successfull result
         pp = ProfilePoint[]
         res = (nothing, pp, :SCAN_BOUND_REACHED)
     elseif ret == :FTOL_REACHED # successfull result
@@ -141,15 +138,14 @@ function get_right_endpoint(
         pp = [ ProfilePoint(optf, loss, optx, ret, nothing) ]
         res = (optf, pp, :BORDER_FOUND_BY_SCAN_TOL)
     else
-        # this part is not normally reached, just for case
-        #throw(ErrorException("No interpretation of the optimization results."))
-        # do not throw
+        # this part is reached in case of :FAILURE return code
+        # we have seen it in case of gradient methods
         pp = ProfilePoint[]
         res = (nothing, pp, :UNKNOWN_STOP)
     end
 
     return res
-end # of bound_right
+end # get_right_endpoint
 
 function get_right_endpoint(
     theta_init::Vector{Float64}, # initial point of parameters
@@ -187,3 +183,5 @@ function get_right_endpoint(
         kwargs... # options for local fitter
     )
 end
+
+# struct OutOfBoundException <: Exception end